pyproject.toml
[project]
name = "instructed-pcg"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "devcomfort", email = "im@devcomfort.me" }
]
dependencies = [
    "absl-py==2.1.0",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "attrs==24.3.0",
    "autobahn==24.4.2",
    "automat==24.8.1",
    "bounded-pool-executor==0.0.3",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "chex==0.1.88",
    "click==8.1.8",
    "cloudpickle==3.1.0",
    "constantly==23.10.4",
    "contourpy==1.3.1",
    "cryptography==44.0.0",
    "cv==1.0.0",
    "cycler==0.12.1",
    "decorator==5.1.1",
    "distrax==0.1.5",
    "dm-tree==0.1.8",
    "docker-pycreds==0.4.0",
    "etils==1.11.0",
    "exceptiongroup==1.2.2",
    "farama-notifications==0.0.4",
    "filelock==3.16.1",
    "flax==0.10.2",
    "fonttools==4.55.3",
    "fsspec==2024.12.0",
    "gast==0.6.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "gym==0.26.2",
    "gym-notices==0.0.8",
    "gymnasium==1.0.0",
    "gymnax==0.0.8",
    "huggingface-hub==0.27.1",
    "humanize==4.11.0",
    "hydra-core==1.3.2",
    "hyperlink==21.0.0",
    "idna==3.10",
    "imageio==2.36.1",
    "importlib-resources==6.5.2",
    "incremental==24.7.2",
    "iniconfig==2.0.0",
    "iterable-io==1.0.0",
    "jax==0.5.0",
    "jax-cuda12-pjrt==0.5.0",
    "jax-cuda12-plugin==0.5.0",
    "jaxlib==0.5.0",
    "kiwisolver==1.4.8",
    "magic-wormhole==0.17.0",
    "markdown-it-py==3.0.0",
    "matplotlib==3.10.0",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.1",
    "msgpack==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.8.3.14",
    "nvidia-cuda-cupti-cu12==12.8.57",
    "nvidia-cuda-nvcc-cu12==12.8.61",
    "nvidia-cuda-runtime-cu12==12.8.57",
    "nvidia-cudnn-cu12==9.7.0.66",
    "nvidia-cufft-cu12==11.3.3.41",
    "nvidia-cusolver-cu12==11.7.2.55",
    "nvidia-cusparse-cu12==12.5.7.53",
    "nvidia-nccl-cu12==2.25.1",
    "nvidia-nvjitlink-cu12==12.8.61",
    "omegaconf==2.3.0",
    "opencv-python==4.10.0.84",
    "opt-einsum==3.4.0",
    "optax==0.2.4",
    "orbax==0.1.9",
    "orbax-checkpoint==0.11.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "polars==1.21.0",
    "pqdm==0.2.0",
    "protobuf==5.29.3",
    "psutil==6.1.1",
    "pyarrow==19.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.10.5",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pynacl==1.5.0",
    "pyopenssl==25.0.0",
    "pyparsing==3.2.1",
    "pytest==8.3.4",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==13.9.4",
    "safetensors==0.5.2",
    "scipy==1.15.1",
    "seaborn==0.13.2",
    "sentry-sdk==2.19.2",
    "service-identity==24.2.0",
    "setproctitle==1.3.4",
    "setuptools==75.8.0",
    "simplejson==3.19.3",
    "six==1.17.0",
    "smmap==5.0.2",
    "spake2==0.9",
    "submitit==1.5.2",
    "tensorboardx==2.6.2.2",
    "tensorflow-probability==0.25.0",
    "tensorstore==0.1.71",
    "tokenizers==0.21.0",
    "tomli==2.2.1",
    "toolz==1.0.0",
    "tqdm==4.67.1",
    "transformers==4.48.0",
    "twisted==24.11.0",
    "txaio==23.1.1",
    "txtorcon==24.8.0",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.3.0",
    "wandb==0.19.2",
    "zipp==3.21.0",
    "zipstream-ng==1.8.0",
    "zope-interface==7.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/instructed_pcg"]
